fixes #34: sanity check obtained page count value
This issue was partially fixed by the robustness changes introduced
for issue #33 (see commit 6c6ecb41278e0911dca80), but this changeset
adds a check specific to issue #34 to verify that the obtained page
count value contains all digits. This keeps any emitted error
messages as close as possible to the source of the problem.