fixes #55 [1 of 2]: repo-create: make repo auto-initialization optional
The ads-github-repo-create tool has a new '-a' (--auto-init) option
that allows the user to control whether or not the newly created
repo will be auto-intialized.

The existing '-b' (--default-branch=NAME) option implies '-a' (--auto-init).

The user is now able to create empty repos on GitHub, which is what
you want when you are starting from an existing Git repo that you
just want to get into GitHub.

See ads-github-repo-create(1) for all the gory details, especially
the "EXAMPLES" and "FAQ" sections. The latter contains a cookbook
for common tasks.