fixes #46 [1 of 2]: fetch-all-upstreams: "fork" vs. "origin"
The ads-github-fetch-all-upstreams tool now asserts the policy that,
in the absense of configuration, the default name of the Git remote
for the user's GitHub fork is "origin" rather than "fork", just as
it would be if the user directly cloned the repo using 'git
clone...'

The git-config(1) keys 'hub.forkremote' and 'hub.upstreamremote' are
checked at the "global" and "system" level to determine if the user
has explicitly asserted a preference for the remote names, and will
use those if present. Otherwise the names "origin" and "upstream"
will be used.

Note that user's who use the 'hub.upstreamremote' key and set its
value to something other than "upstream" will actually end up with
three remotes rather than two. The logical upstream remote will have
one named according to the user's declared preference, and one named
"upstream". This is because the name "upstream" is currently
baked-in at a fairly low level, and making that more general was out
of scope for the current issue (which was primarily aimed at the
"fork" vs. "origin" problem).