add file (program): contrilb/curl-gh
Base-level capability for invoking curl(1) on the command line in a
way that is similar to how the various 'ads-github-tools' programs
do it. As the base-level program, this one is "furthest away" from
that, in the sense that it is closest to the default behavior of
curl; curl's default output is not altered, and no temporary files
are explicitly written. See comment at the top of the source code
for details.

Two additional programs ('curl-gh-fout' and 'curl-gh-fout-q') will
be committed momentarily; they "layer on" additional options that
get successively closer to the actual behavior of the
'ads-github-tools' programs.