add file (program): contrilb/curl-gh-fout-q
Wraps 'curl-gh-fout' to "layer on" the behavior of writing the HTTP
resposne code to stdout. The default curl(1) progress output is
suppressed, but not error or warning messages.

Of the 'curl-gh*' set of programs, this one most closely mimics the
behavior of the various 'ads-github-tools' programs. The the HTTP
response body will be written to the user-specified filepath, and
the HTTP response code will be emitted on stdout.