add file: autotools/ads_PROG_CARGO.m4
This macro is intended to be used in configure.ac. It locates the
'cargo' program at configure time, and accepts an optional parameter
to allow a minimum required version to be specified.

If no specific required version is specified, then the default used
is 1.0.0, which is arbitrary, but old enough that it should work for
anything found in the wild. Note that the cargo 1.0.0. release was
made on 2015-05-15 (more than five years ago).

I should also note here that I have not tested this macro against
versions of cargo older than 1.42.1. I don't actually know if the
'--version' option was even present in those old verions.

The version comparision honors /most/ of the SemVer 2.0 spec:

    - The M.m.p ("major", "minor", and "patch") field comparisons
      are fully implemented.

    - The "pre-release" field is partially implemented. If two
      version strings have the same M.m.p values, but only one of
      them also has a "pre-release" field present, then the one with
      the "pre-release" field is treated as the older.

      The SemVer 2.0 spec further defines comparisons between
      subfields within the "pre-release" field. Those comparisons
      are not implemented. They do not seem all that important for
      our use case; we can revisit this decision, if needed.