add file: autotools/ads_PROG_CARGO-dev-helper
This is the tool use when initiall implementing the
ads_PROG_CARGO.m4 macro. It is where the somewhat elaborate regexen
were tested against both good and bad data (also contained within
the script, with some examples coming directly from the SemVer 2.0
spec).

If we ever need to adjust the regexen used by the macro, a good
approach would be to get them working in a test bed version of this
script first.