issue #59: add program: parse-netrc
The 'parse-netrc' program is a Rust-based command line utility that
parses the user's ~/.netrc file to find the first record that
matches a user-provided hostname. When found, the program emits the
record's "login" (username) field value to stdout.

--------------------------------->8---------------------------------
    $ parse-netrc --help
    usage: parse-netrc { -h | --help }
      or:  parse-netrc { -V | --version }
      or:  parse-netrc [OPTION...] { -u USER | --user=USER } [--] HOSTNAME

    Extract and print fields from matching netrc record, if any.

    Mandatory arguments to long options are mandatory for short options too.

      -h, --help        Print this help message on stdout
      -V, --version     Print the version of the program on stdout
      -u, --user=USER   Require match of USER in matched netrc record
      -v, --verbose     Print program progress messages on stderr. Specify multiple
                          times to increase verbosity: info, debug, and tracing
          --            Signals the end of options and disables further options processing.
                          Any remaining argument(s) will be interpretted as a hostname

    Report bugs to Alan D. Salewski <ads@salewski.email>.
---------------------------------8<---------------------------------

In the context of the 'ads-github-tools', this provides us with a
client-side mechanism to determine what the in-effect GitHub
username is without contacting the GitHub API (as
'ads-github-whoami' does). This is necessary for our cache
implementation, as the local Unix user account may have more than
one GitHub account; we want to keep the per-GitHub-user cache
information stored in isolation. The who purpose of the cache is to
/avoid/ unnecessary network calls, so relying on the GitHub API for
this pupose is a non-starter.

Note that the underlying 'netrc' (third-party) crate is not able to
parse netrc files that contain comments. See comments in the
'parse-netrc' program source for other limitations and future plans
for addressing that sort of thing.

This is a rather large changeset mainly because this is the first
Rust-based program being integrated into the 'ads-github-tools'
project. We add all the necessary scaffolding to integrate the Cargo
build tool with our existing GNU Autotools based build.

Note that this does not preclude using the 'cargo' command
directly. Once you have your tree configured, you can freely
intermix (more or less) 'make' and 'cargo' invocations as the
situation warrants.

Files that influence the Rust-based build:

    $(top_srcdir)/.cargo/config.toml
    $(top_srcdir)/src/main/rust/bin/parse-netrc/.cargo/config.toml
    $(top_srcdir)/src/main/rust/bin/parse-netrc/Cargo.toml

Note, too, that all third-party Rust library deps (both direct and
transitive) are "vendored" under the directory:

    $(top_srcdir)/src/third-party/cargo-vendored/

The Cargo builds performed through the Makefile interface are
configured to reference those artifacts only, and never anything
from the Internet.

There's a write-up for how most of this hangs together in the long
comments of the file:

    $(top_srcdir)/src/main/rust/bin/Makefile.am

Why there? Because that is the parent directory of every Cargo
"package" in the project tree. It is the border line between "GNU
Autotools World" and "Rust/Cargo World", so it seems like a good
place to put a visitors' center. It seems weird at the moment, in
part, because we happen to have only one Cargo "package". In the
future, those notes may find their way into the top-level 'HACKING'
file, or similar.