issue #59: add program: ads-github-cache
The 'ads-github-cache' program implements a user-specific cache of
GitHub objects, and provides a command line interface through which
the cache can be manipulated.

At the moment, the only API resource path that is officially
supported is:

    /user/repos

but that has the benefit of being the most intense point of pain in
day-to-day usage. For example, the author often avoids cloning repos
via '*-fetch-all-upstreams -vcu' because it takes so long (about a
minute) to gather the necessary data via the GitHub API before
performing the actual operation.

Support for retrieving the data at the above path (and others) still
needs to be incorporated into the various utilities, but now we have
a foundation upon which to build. The difference between non-cached
and cached data retrieval is significant.

Consider the following. The first is the amount of time it takes for
the author to retrieve all of the data across 18 requests (at 100
items per page, which is the GitHub-allowed maximum, and the
built-in default of ads-github-cache). This sample is actually
pretty fast for this particular chunk of data. In real-world use
cases, this request tends to take nearly twice as long as it took
here (it's 05:00 EDT, so it seems this might be the sweet spot for
accessing the GitHub servers  :-)

    $ ads-github-cache --clear-all
    $ time ads-github-cache --get '/user/repos' > /dev/null

    real    0m33.086s
    user    0m1.411s
    sys     0m0.740s

Here we follow-up with the same action, expecting it to be faster
because all of the data items are already present in the local
cache. This is still a network-based operation. The '--get' request
uses the "ETag" values retrieved with the original requests to
request only updates from the server. So there is a round-trip
network call to do the asking, but (in this case) no large chunks of
data need to be gathered on the server side or sent across the
network for us to save:

    $ time ads-github-cache --get '/user/repos' > /dev/null

    real    0m19.297s
    user    0m0.974s
    sys     0m0.340s

So that's an improvement, and that ~20 seconds number is about what
we see regularly for this "get through cache" operation.

Here's what it looks like when pulling that same data directly from
the cache, without allowing for network access:

    $ time ads-github-cache --get-cached '/user/repos' > /dev/null

    real    0m0.272s
    user    0m0.252s
    sys     0m0.046s

Boom. That's what /I'm/ talkin' 'bout. And this is in a (somewhat
crazy) shell script; we could make it go even faster, but at this
speed there's not (currently) much point to doing so.

--------------------------------->8---------------------------------
    $ ads-github-cache --help
    usage: ads-github-cache { -h | --help }
      or:  ads-github-cache { -V | --version }
      or:  ads-github-cache [OPTION...] --update [--] [URL_OR_PATH...]
      or:  ads-github-cache [OPTION...] { --get | --get-cached } [--] URL_OR_PATH
      or:  ads-github-cache [OPTION...] --clear  [--] [URL_OR_PATH...]
      or:  ads-github-cache [OPTION...] --clear-all

    Manage a user-specific cache of GitHub v3 API responses.

    Mandatory arguments to long options are mandatory for short options too.

      -h, --help        Print this help message on stdout
      -V, --version     Print the version of the program on stdout
          --clear       Remove the specified URLs or paths from the cache
          --clear-all   Remove all cached entries, for all cached URLs or paths
          --get         Obtain content of specfied URL or path through the cache
          --get-cached  Like '--get', but error out if item is not already present in cache.
                          Avoids updating the cache to obtain the item. Think "offline".
          --update      Update the cache entry for the specified URLs or paths, or all
      -v, --verbose     Print program progress messages on stderr. Specify multiple
                          times to increase verbosity: info, debug, and tracing (set -x)
          --            Signals the end of options and disables further options processing.
                          Any remaining argument(s) will be interpretted as a repo name

    Report bugs to Alan D. Salewski <ads@salewski.email>.
---------------------------------8<---------------------------------