fixes #61: fetch-all-upstreams: leverage cache for /user/repos data
WORKFLOW BREAKAGE (the good kind): The default behavior of the
program has changed to expect that the cached data is present and
available. By default, it attempts to access the cached data in
'offline' mode, which is the fastest. This means use user must have
successfully run 'ads-github-cache --update' or an error will be
emitted due to the cached data not being available.

The pain of this workflow breakage is considered to be very low, and
the benefit of the change very high.

There is a new '--cache-mode=MODE' command line option that allows
the user to select which caching behavior is desired: 'offline' (the
default), 'online', or 'none'.

If you just want to get the old, slow behavior, say '--cache-mode=none'.

See ads-github-fetch-all-upstreams(1) for the full scoop.

The author sees the following timing differences:

        avoid cache: ~58 seconds    (legacy behavior)
     cache (online): ~19 seconds
    cache (offline): < 1 second     (new default behavior)