fixes #64: repo-create: JSON-escape string values
The 'ads-github-repo-create' tool was not properly escaping JSON
strings during new repo creation. When users supplied command line
parameters that contained string values that would need such
escaping in a JSON document, the program submitted JSON to the
GitHub API that was not well-formed. That resulted in an HTTP 400
response, with the response body:

    {"message":"Problems parsing JSON","documentation_url":"https://docs.github.com/rest/reference/repos#create-a-repository-for-the-authenticated-user"}

The program has been fixed to properly escape those string values.