fixes #64: repo-create: JSON-escape string values (no extra newline)
The fix for this issue earlier today in commit 7d84e3c8a084 properly
escaped the JSON strings, but also inadvertently introduced a
newline ('\n') character into the string value that was escaped.

The input string "foo" was coming out as "foo\n".

Not only does that count as data corruption, but it also caused the
JSON document to be rejected by the GitHub API with the error:

    "description control characters are not allowed"

The program has been fixed to avoid that issue, while still properly
escaping the input strings.