[1 of 2] build: integrate cargo verbosity with the automake silencing rules
Also, because the output of 'cargo --verbose' is voluminous enough
to make normal debugging more (not less) difficult, we do not want
it to be the default. We therefore enable the AM_SILENT_RULES by
default in configure.ac. See notes in the comments there for how to
override override the default either configure time or at build
time.

Note that this is just a minimal integration with the Automake
silencing rules. Much more can be done in the future to improve the
integration and make the difference between silent/non-silent output
more dramatic.