[1 of 3] add file: configure-time.rs.in
We're using this as something of a rough analogue of the 'config.h'
file you might see on C-based projects that use the GNU Autotools.

The idea is that we want to isolate, to the extent possible, the
configure- and build-time filtering done by our Rust program to this
single file. Right now it lives directly in the source tree of the
'parse-netrc' program, because that is the only Rust program
currently in our project source tree. If we were to add other
programs, however, than we would want to move that file into a
project-specific crate shared amongst all of them.

This change only introduces the *.in template file and causes it to
generate the 'configure-time.rs' file at build time. Additional
changes will be required to use the generated file from the
'parse-netrc' program.