parse-netrc: move src/main.rs to src/bin/main.rs
Though the previous approach is consistent with Rust naming
conventions for simple projects, this new arrangement is intended to
make more clear the separation between the library code and that of
the 'parse-netrc' binary crate (since they live in the same Cargo
"package").

The 'lib.rs' file controls the structure and visibility of any
modules. The 'src/bin/main.rs' file will only use project modules
via the public interface of the library (it will not source any of
the other *.rs source files directly).