[1 of 2]: configure.ac: list all static rust (.rs) source files in AC_CONFIG_LINKS
The 'make distcheck' target was failing because Cargo was
complaining that it could not find the static Rust (.rs) source
files. There were actually two problems:

    1. The static rust source files were no longer being included in
       the source tarball.

    2. VPATH builds were looking for all of the rust source files in
       the build tree; only the generated rust source file(s) were
       there.

Listing all of the static rust source files in AC_CONFIG_LINKS
solves both problems.

Background
----------
We generate at least one of our Rust source files for the
'parse-netrc' program, which gets emitted in the build
directory. During a VPATH build, that will be different than
the (possibly read-only) source directory.

We rely on the Cargo build tool to build the 'parse-netrc' program,
and Cargo wants all of the Rust source files under consideration to
be located beneath a single directory. Since we generate at least
one Rust source file, that directory used by Cargo must be in the
build tree.

Therefore, we now use AC_CONFIG_LINKS to make the static Rust source
files appear to be located in the build tree. In most cases, the
mechanism used for this will by symlinks back into the source tree,
but could involve some other workalike mechanism depending on the
target system. See the Autoconf manual for details.

Signed-off-by: Alan D. Salewski <ads@salewski.email>