fixes #78: invoke 'git-hub' rather than 'git hub'
In earlier versions of the 'ads-github-tools' project, the
'configure' script did look for a program named 'git-hub' on $PATH
at configure time, but at runtime the invocation used was that for
Git subcommands:

    git hub ...

That meant the location of the 'git-hub' program found at runtime
was dependent upon the user's current $PATH (and whatever other
rules 'git' uses internally) rather than the $PATH as it was set at
configure time.

This behavior has been changed so that the program is invoked
directly with the name 'git-hub' rather than as a 'git' subcommand.

The only program affected by this change is
'ads-github-fetch-all-upstreams'. The change does not result in any
change to user-visible behavior in the common case. Where one could
conceivably observe different behavior is in the unlikely event that
all of the following are true:

    1. the user both has multiple 'git-hub' programs installed;

    2. the user has different configure-time and runtime $PATH that
       affect which 'git-hub' program was found in each scenario;

    3. and those differing 'git-hub' programs themselves exhibit
       different behavior for the 'clone' operation