fixes #74: repo-create: avoid choking on 'visibility' field
---- meta: start ---------------->8---------------------------------
    Meta: The fix for issue #74[0] was actually fixed as a
    side-effect of the work done[1] to address issue #70[2] in
    commit:

        006e2b28a5303e879e5777b3ffb5b3fe79be830b

    The current commit merely updates the 'NEWS' file to note the
    inclusion of the fix.
---- meta: end ------------------8<---------------------------------

The API response emitted by the upstream GitHub v3 API for the endpoint:

    /repos/:owner/:repo

changed regarding the 'visiblity' field. That change broke the
verification code in the 'ads-github-repo-create' tool that was
performing its checks based on the old behavior.

The impact of the breakage for users was that full automated
verification of the newly created repo was not possible without the
fix provided with the current release of the 'ads-github-tools'
(anticipated to version 0.3.4). The repo got created as
intended (assuming there were no other problems); it was only the
automated verification checking that choked. Nevertheless, manual
verification was still necessary by poking around the GitHub web UI,
which was tedious to do even for a single newly created repo.

Specifically, when the 'ads-github-repo-create' program was
originally implemented a few years ago, support for the 'visibility'
field was a GitHub "opt-in only" experimental feature, and the field
would either be omitted or null in the response when not opting
in. Fast-forward to the present, and the field seems to be present
in the response always. The program's verification code has
therefore been updated accordingly.

[0] "repo-create: verification choking on 'visibility' field when 'private' rather than '-'"
    https://github.com/salewski/ads-github-tools/issues/74

[1] https://github.com/salewski/ads-github-tools/commit/006e2b28a5303e879e5777b3ffb5b3fe79be830b

[2] "repo-create: allow creation of a repo in a GitHub Organization"
    https://github.com/salewski/ads-github-tools/issues/70

Signed-off-by: Alan D. Salewski <ads@salewski.email>