[3/6]: configure.ac: check for sleep(1) utility
In an upcoming commit, the 'ads-github-cache' program will be making
use of the sleep(1) utility, so we add a configure-time check to
ensure one exists and learn of its location.

See comments in code for note on the deficiency of this check to
account for the real need of an implementation that supports
sub-second sleep time specifications (such as those from the BSDs
and the GNU coreutils). This is good enough for now, however, since
it will work on all known platforms where the ads-github-tools are
actually run currently.

Signed-off-by: Alan D. Salewski <ads@salewski.email>