[5/6]: ads-github-cache: update collections concurrently in subshells
The existing functionality of the f_get_one_collection_or_die()
function is to invoke the f_perform_do_get_one_or_die() function to
update (if needed) the locally cached data for each page of a "paged
collection". Historically, such updates were done sequentially. The
current change modifies that behavior to perform the updates for
each individual page in a subshell, in the background.

This new behavior creates the opportunity for all of the requests to
make progress concurrently, since much time involved in such
operations is waiting for the remote GitHub v3 API to respond.

Anecdotal testing shows a significant real-world improvement. For
example, the author's invocations of:

    $ ads-github-cache --verbose --update

drop from more than two minutes (with the previous behavior) down to
between nine and eleven seconds.

Signed-off-by: Alan D. Salewski <ads@salewski.email>