[2/6]: cache: retool variables, functions to allow running concurrently
Retool ads-github-cache.in to allow select shell functions to run
concurrently (in subshell processes).

Note that this change does not actually attempt to perform
operations concurrently; it just makes changes that will allow for
that (in a future commit). This basically comes down to removing
assumptions that were baked into some functions regarding access to
certain file paths in the program's temporary directory -- usage of
such files now uses file names that are unique not just to the
invocation of the program, but also to the subshell process that is
running the function.

Since these changes imply functions will be running in subprocesses,
we also needed to export a large number of our variables. (An
/uncomfortably/ large number of them, in fact.)

Signed-off-by: Alan D. Salewski <ads@salewski.email>