issue #81: [3/4]: cache: limit background procs based on cpu count
When retrieving pages of data (from the remote GitHub site) for a
path that represents a "paged collection", the ads-github-cache
program now limits the number of subshell subprocesses it will spawn
at any one moment (rather than just spawning a subshell for all
pages all at the same time).

This change limits the number based on two criteria:

    1. The number of available CPUs found. The limit imposed
       is (arbitrarily) twice this number, recognizing that even on
       a single-CPU machine it is likely more efficient to have at
       least two concurrent HTTP requests to the GitHub API (since
       there will be a fair amount of time just waiting for the
       response to come back).

       Note that the "number of available CPUs" is determined using
       the recently added 'ads-github-nproc' program.

    2. A hard cap of 20 outstanding subshell processes. This is just
       an arbitrarily chosen number that is intended to keep the
       program from making an overly-large number of concurrent
       requests to the GitHub API, which would otherwise be the case
       with just (1) above on a machine with a large number of
       CPUs.

Note that a future commit will allow the user to specify the number
directly, so in the end there will be three criteria controlling
this behavior.

Signed-off-by: Alan D. Salewski <ads@salewski.email>