issue #81: [4/4]: cache: add -j NUM (--jobs=NUM) option
The 'ads-github-cache' program now accepts a '-j NUM' (--jobs=NUM)
command line option that allows the user to specify the maximum
number of requests to the GitHub API that can be created
simultaneously.

By default (without the command line option), the number is twice
the number of CPUs found to be available at runtime, but with a hard
cap of 20 (to avoid making an excessive number of requests on
machines that happen to have a larger number of CPUs).

If provided, the user-specified number will be used unconditionally,
without any cap (assuming the machine has enough resources
available). Please use this feature responsibly.

Note that values of NUM must be an integer greater than zero (the
program will reject other values and exit with an error status).

If the value 1 is provided for NUM, the effect will be to eliminate
concurrent requests.

Signed-off-by: Alan D. Salewski <ads@salewski.email>