fixes #84: doc: HACKING: update to reflect use of newer versions of the GNU Autotools
The build process itself was updated back in issue #79 (part of the
'ads-github-tools-0.3.4' release), but those earlier changes
neglected to update the documentation in the 'HACKING' file.

The docs have also been improved by noting where source tarballs may
be obtained, which is helpful for side-stepping the need to even
have the GNU Autotools installed.

Signed-off-by: Alan D. Salewski <ads@salewski.email>