issue #85: cache: serialize data emit
The concurrency features added to the 'ads-github-cache' tool in the
ads-github-tools-0.3.4 release (see issues #80[0] and #81[1])
inadvertently broke some operations that emitted cached data. Data
was not only concurrently fetched for updating the cache, but was
sometimes emitted concurrently, as well, causing the output to be
garbled (mainly in the form of JSON that was no longer well-formed
due to being intermixed from multiple emitting background
processes).

With the fix in this changeset, the data is still fetched
concurrently (from the upstream GitHub v3 API endpoints), but
emitting the data from cached objects now happens sequentially as
one would expect. This is mainly relevant for paths that represent
"paged collections", where multiple pages of data are available.

[0] issue #80: "cache (new feature): update multiple pages for a resource path concurrently"
    https://github.com/salewski/ads-github-tools/issues/80

[1] issue #81: "cache (new feature): add '-j' (--jobs=NUM) to control concurrency"
    https://github.com/salewski/ads-github-tools/issues/81

Signed-off-by: Alan D. Salewski <ads@salewski.email>